refactor fetch userData out of server actions file into separate helper file
