+++ src/app/(auth)/dashboard/[...cardId]/not-found.tsx
@@ -2,7 +2,7 @@ import CommonHero from '@/components/common/CommonHero';
 import { Hero } from '@/types/CommonHero.type';
 import Dashboard404 from '@/components/common/Dashboard404';
 import Header from '@/components/common/Header';
-import { getUserData } from '@/app/actions/userDataActions';
+import { getUserData } from '@/app/utils/fetchHelper';
 import { UserData } from '@/types/User.type';
 
 export default async function NotFound() {
