+++ src/app/(auth)/help-support/page.tsx
@@ -5,7 +5,8 @@ import BackToTop from '@/components/common/BackToTop';
 import Footer from '@/components/common/Footer';
 import Accordion from '@/components/user-portal/FAQ';
 import { isAuthenticated } from '@/app/utils/auth/amplify-utils';
-import addUserPostSignIn, { getUserData } from '@/app/actions/userDataActions';
+import addUserPostSignIn from '@/app/actions/userDataActions';
+import { getUserData } from '@/app/utils/fetchHelper';
 import { invitationData, UserData } from '@/types/User.type';
 import { redirect } from 'next/navigation';
