+++ src/app/privacy-policy/page.tsx
@@ -5,7 +5,7 @@ import Header from '@/components/common/Header';
 import { SEO_CONFIG } from '@/utils/seoConfig';
 import PrivacyPolicy from '@/components/terms/PrivacyPolicy';
 import { Metadata } from 'next';
-import { getUserData } from '@/app/actions/userDataActions';
+import { getUserData } from '@/app/utils/fetchHelper';
 import { UserData } from '@/types/User.type';
 
 export const metadata: Metadata = {
