+++ src/app/sign-up/page.tsx
@@ -4,7 +4,7 @@ import Header from '@/components/common/Header';
 import SignUpForm from '@/components/sign-up/SignUpForm';
 import { SEO_CONFIG } from '@/utils/seoConfig';
 import { Metadata } from 'next';
-import { getUserData } from '@/app/actions/userDataActions';
+import { getUserData } from '@/app/utils/fetchHelper';
 import { UserData } from '@/types/User.type';
 import { Hero } from '@/types/CommonHero.type';
